Moved out service and hook from useMaintenanceBanner to the specific files
